pyproject.toml
[project]
name = "tradingbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ccxt>=4.5.12",
    "joblib>=1.5.2",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "scikit-learn>=1.7.2",
    "scikit-optimize>=0.10.2",
    "streamlit>=1.50.0",
    "ta>=0.11.0",
]
